separate client and server classes
