+++ examples/client_with_routing.py
@@ -9,7 +9,7 @@ from rsocket.extensions.mimetypes import WellKnownMimeTypes
 from rsocket.payload import Payload
 from rsocket.queue_response_stream import QueueResponseStream
 from rsocket.routing.helpers import route, composite, authenticate_simple
-from rsocket.rsocket import RSocket
+from rsocket.rsocket_client import RSocketClient
 
 
 class RequestChannel(QueueResponseStream, Subscriber):
@@ -74,9 +74,8 @@ class StreamSubscriber(Subscriber):
 
 
 async def communicate(reader, writer):
-    async with RSocket(reader, writer,
-                       server=False,
-                       metadata_encoding=WellKnownMimeTypes.MESSAGE_RSOCKET_COMPOSITE_METADATA) as socket:
+    async with RSocketClient(reader, writer,
+                             metadata_encoding=WellKnownMimeTypes.MESSAGE_RSOCKET_COMPOSITE_METADATA) as socket:
         await test_request_response(socket)
         await test_stream(socket)
         await test_slow_stream(socket)
@@ -84,14 +83,14 @@ async def communicate(reader, writer):
         await test_stream_invalid_login(socket)
 
 
-async def test_request_response(socket: RSocket):
+async def test_request_response(socket: RSocketClient):
     payload = Payload(b'The quick brown fox',
                       composite(route('single_request'), authenticate_simple('user', '12345')))
 
     await socket.request_response(payload)
 
 
-async def test_channel(socket: RSocket):
+async def test_channel(socket: RSocketClient):
     channel_completion_event = Event()
     requester_completion_event = Event()
     channel_payload = Payload(b'The quick brown fox',
@@ -104,21 +103,21 @@ async def test_channel(socket: RSocket):
     await requester_completion_event.wait()
 
 
-async def test_stream_invalid_login(socket: RSocket):
+async def test_stream_invalid_login(socket: RSocketClient):
     payload = Payload(b'The quick brown fox', composite(route('stream'), authenticate_simple('user', 'wrong_password')))
     completion_event = Event()
     socket.request_stream(payload).limit_rate(1).subscribe(StreamSubscriber(completion_event))
     await completion_event.wait()
 
 
-async def test_stream(socket: RSocket):
+async def test_stream(socket: RSocketClient):
     payload = Payload(b'The quick brown fox', composite(route('stream'), authenticate_simple('user', '12345')))
     completion_event = Event()
     socket.request_stream(payload).subscribe(StreamSubscriber(completion_event))
     await completion_event.wait()
 
 
-async def test_slow_stream(socket: RSocket):
+async def test_slow_stream(socket: RSocketClient):
     payload = Payload(b'The quick brown fox', composite(route('stream-slow'), authenticate_simple('user', '12345')))
     completion_event = Event()
     socket.request_stream(payload).subscribe(StreamSubscriber(completion_event))
