+++ tests/rsocket/conftest.py
@@ -1,20 +1,22 @@
 import asyncio
+
 import pytest
 
-from rsocket.rsocket import RSocket
+from rsocket.rsocket_client import RSocketClient
+from rsocket.rsocket_server import RSocketServer
 
 
 @pytest.fixture
 def pipe(unused_tcp_port, event_loop):
     def session(reader, writer):
         nonlocal server
-        server = RSocket(reader, writer)
+        server = RSocketServer(reader, writer)
 
     async def start():
         nonlocal service, client
         service = await asyncio.start_server(session, host, port)
         connection = await asyncio.open_connection(host, port)
-        client = RSocket(*connection, server=False)
+        client = RSocketClient(*connection)
 
     async def finish():
         service.close()
